+++ lib/config/init.js
@@ -2,6 +2,7 @@ import fs from 'fs';
 import nconf from 'nconf';
 import path from 'path';
 
+
 // Config
 nconf.use('memory');
 nconf.argv().env();
