+++ lib/config/sentry.js
@@ -1,6 +1,7 @@
 import nconf from 'nconf';
 import raven from 'raven';
 
+
 let client = {
   captureError: (err) => console.log(err)
 };
