+++ lib/dota2.js
@@ -6,10 +6,11 @@ import _request from 'request';
 import sentry from './config/sentry';
 import { getOrdinal, numberWithCommas, secondDec, toTitleCase } from './helpers';
 
+
 const request = Promise.promisify(_request);
 
 const help_text = `Dota 2 help commands
-Accepted positions are **Mid**, **Off**, **Safe**, or **Jungle**.
+Accepted positions are **Mid**, **Off**, **Safe**, and **Jungle**.
 
 **\`!dota2 best\`** \`position\`
     Get the top 10 Heroes for a specific position
