+++ lib/fun/comics.js
@@ -5,10 +5,10 @@ import R from 'ramda';
 
 import sentry from './config/sentry';
 
+
 const request = Promise.promisify(_request);
 
-const help_text = `Comics commands
-Get random web comic strips
+const help_text = `Get random web comic strips
 
 **\`!comic random\`**
     Returns a random comic from any of the artists
@@ -19,7 +19,7 @@ Get random web comic strips
 **\`!comic cah\`** or **\`!comic cyanide and happiness\`**
     Returns a Cyanide and Happimess comic
 **\`!comic css\`** or **\`!comic chainsawsuit\`**
-    Returns a chainsawsuit comic
+    Returns a Chainsawsuit comic
 **\`!comic dhd\`** or **\`!comic dog house diaries\`**
     Returns a Dog House Diaries comic
 **\`!comic smbc\`** or ** \`!comic saturday morning breakfast cereal\`**
