+++ lib/fun/meme.js
@@ -1,30 +1,12 @@
 import Imgflipper from 'imgflipper';
-import gizoogle from 'gizoogle';
 import nconf from 'nconf';
 
 import sentry from './config/sentry';
-import { drama as _drama, emoji as _emoji, memes } from '../data';
+
+import { memes } from '../data';
+
 
 const imgflipper = new Imgflipper(nconf.get('IMGFLIP_USERNAME'), nconf.get('IMGFLIP_PASSWORD'));
-
-
-function drama(bot, msg, suffix) {
-  let rand = Math.floor(Math.random() * _drama.length);
-  if (suffix && suffix >= 0 && suffix <= (_drama.length - 1)) {
-    bot.sendMessage(msg.channel, _drama[suffix]);
-  } else {
-    bot.sendMessage(msg.channel, _drama[rand]);
-  }
-}
-
-function emoji(bot, msg, suffix) {
-  let rand = Math.floor(Math.random() * _emoji.length);
-  if (suffix && suffix >= 0 && suffix <= (_emoji.length - 1)) {
-    bot.sendMessage(msg.channel, _emoji[suffix]);
-  } else {
-    bot.sendMessage(msg.channel, _emoji[rand]);
-  }
-}
 
 function meme(bot, msg, suffix) {
   if (!nconf.get('IMGFLIP_USERNAME') || !nconf.get('IMGFLIP_PASSWORD')) {
@@ -53,22 +35,7 @@ function meme(bot, msg, suffix) {
   });
 }
 
-function snoopify(bot, msg, suffix) {
-  if (!suffix) {
-    bot.sendMessage(msg.channel, 'Usage: **`!snoopify`** `sentence`');
-    return;
-  }
-  gizoogle.string(suffix, (err, translation) => {
-    if (err) sentry.captureError(err);
-    bot.sendMessage(msg.channel, translation);
-  });
-}
-
 export default {
-  drama,
-  emoji,
-  emojis: emoji,
   meme,
-  memes: meme,
-  snoopify
+  memes: meme
 };
