+++ lib/fun/translate.js
@@ -1,11 +1,24 @@
 import Promise from 'bluebird';
 import cheerio from 'cheerio';
+import gizoogle from 'gizoogle';
 import _request from 'request';
 import R from 'ramda';
 
 import sentry from './config/sentry';
 
+
 const request = Promise.promisify(_request);
+
+function snoop(bot, msg, suffix) {
+  if (!suffix) {
+    bot.sendMessage(msg.channel, 'Usage: **`!snoopify`** `sentence`');
+    return;
+  }
+  gizoogle.string(suffix, (err, translation) => {
+    if (err) sentry.captureError(err);
+    bot.sendMessage(msg.channel, translation);
+  });
+}
 
 function yoda(bot, msg, phrase) {
   if (!phrase) {
@@ -34,6 +47,8 @@ function yoda(bot, msg, phrase) {
 }
 
 export default {
+  snoop,
+  snoopify: snoop,
   yoda: yoda,
   yodaify: yoda
 };
