+++ lib/help/index.js
@@ -2,6 +2,7 @@ import english from './english';
 import french from './french';
 import meme from './meme';
 
+
 export default {
   aide: (bot, msg, suffix) => {
     suffix = suffix.toLowerCase();
