+++ lib/info.js
@@ -3,8 +3,9 @@ import R from 'ramda';
 import _request from 'request';
 
 import sentry from './config/sentry';
-const request = Promise.promisify(_request);
 
+
+const request = Promise.promisify(_request);
 
 function avatar(bot, msg, suffix) {
   if (msg.mentions.length === 0) {
@@ -13,27 +14,12 @@ function avatar(bot, msg, suffix) {
     } else {
       bot.sendMessage(msg.channel, `Your avatar:\n${msg.author.avatarURL}`);
     }
-    return;
   }
   let msg_array = R.map(user => {
     if (user.avatarURL === null) return `${user.username} is naked.`;
     return `${user.username}\'s avatar:\n${user.avatarURL}`;
   }, msg.mentions);
   bot.sendMessage(msg.channel, msg_array);
-}
-
-function joinServer(bot, msg, suffix) {
-  if (!suffix) {
-    bot.sendMessage(msg.channel, 'Usage: **`!join`** `invitation link`');
-    return;
-  }
-  bot.joinServer(suffix, (err, server) => {
-    if (err) {
-      bot.sendMessage(msg.channel, `Failed to join: ${err}`);
-    } else {
-      bot.sendMessage(msg.channel, `Successfully joined ${server}`);
-    }
-  });
 }
 
 function serverinfo(bot, msg) {
@@ -145,10 +131,6 @@ export default {
   avatar,
   changelog: version,
   'change-log': version,
-  invite: joinServer,
-  join: joinServer,
-  joinserver: joinServer,
-  'join-server': joinServer,
   newfeatures: version,
   'new-features': version,
   serverinfo,
@@ -158,5 +140,6 @@ export default {
   stats: servers,
   userinfo,
   uptime,
-  version
+  version,
+  whois: userinfo
 };
