+++ lib/leagueoflegends/championgg.js
@@ -38,7 +38,6 @@ const positions = {
   supp: 'Support'
 };
 
-
 // Setup and makes request to Champion.GG API
 function _makeRequest(options) {
   let default_options = {
