+++ lib/leagueoflegends/index.js
@@ -2,6 +2,7 @@ import { help_text } from './help';
 import { bans, best, counters, items, skills } from './championgg';
 import { serverStatus, matchDetails } from './riot';
 
+
 export default {
   lol: (bot, msg, suffix) => {
     let command = suffix.toLowerCase().split(' ')[0];
@@ -18,6 +19,7 @@ export default {
     if (command === 'match') return matchDetails(bot, msg, suffix);
     if (command === 'skill') return skills(bot, msg, suffix);
     if (command === 'skills') return skills(bot, msg, suffix);
+    if (command === 'server') return serverStatus(bot, msg, suffix);
     if (command === 'server-status') return serverStatus(bot, msg, suffix);
     if (command === 'servers') return serverStatus(bot, msg, suffix);
     if (command === 'serverstatus') return serverStatus(bot, msg, suffix);
