+++ lib/leagueoflegends/riot.js
@@ -22,7 +22,6 @@ const NotFoundError = SuperError.subclass('NotFoundError', function(msg) {
 // List of regions
 const regions = ['br', 'eune', 'euw', 'kr', 'lan', 'las', 'na', 'oce', 'pbe', 'ru', 'tr'];
 
-
 // Setup and makes request to Riot API
 function _makeRequest(url) {
   let options = {
@@ -38,7 +37,6 @@ function _makeRequest(url) {
     })
     .then(R.prop('body'));
 }
-
 
 export function serverStatus(bot, msg) {
   let options = {
