+++ lib/other.js
@@ -1,5 +1,6 @@
 import path from 'path';
 
+
 export default {
   ayylmao: (bot, msg) => {
     bot.sendMessage(msg.channel, 'http://i.imgur.com/m7NaGVx.gif');
