+++ lib/useful/appearin.js
@@ -1,7 +1,7 @@
 import R from 'ramda';
 
 
-function appearin(bot, msg) {
+function appearin(bot, msg, suffix) {
   let id = Math.random().toString().replace('.', '').slice(0, 10);
   let url = `https://appear.in/${id}`;
