+++ lib/useful/gif.js
@@ -4,9 +4,15 @@ import R from 'ramda';
 
 import sentry from './config/sentry';
 
+
 const request = Promise.promisify(_request);
 
-function giphy(bot, msg, tags) {
+function giphy(bot, msg, suffix) {
+  if (!suffix) {
+    bot.sendMessage(msg.channel, 'Usage: **`!giphy`** `gif tags`');
+    return;
+  }
+
   // limit=1 will only return 1 gif
   let options = {
     url: 'http://api.giphy.com/v1/gifs/search',
@@ -18,7 +24,7 @@ function giphy(bot, msg, tags) {
     },
     json: true
   };
-  if (tags) options.qs.q = tags.join('+');
+  if (suffix) options.qs.q = suffix.join('+');
 
   request(options)
     .then(R.prop('body'))
@@ -26,7 +32,7 @@ function giphy(bot, msg, tags) {
       if (body.data.length) {
         bot.sendMessage(msg.channel, `http://media.giphy.com/media/${body.data[0].id}/giphy.gif`);
       } else {
-        bot.sendMessage(msg.channel, `I couldn\'t find a gif for: ${tags}`);
+        bot.sendMessage(msg.channel, `I couldn\'t find a gif for: ${suffix}`);
       }
     })
     .catch(err => {
@@ -36,12 +42,5 @@ function giphy(bot, msg, tags) {
 }
 
 export default {
-  gif: (bot, msg, suffix) => {
-    if (!suffix) {
-      bot.sendMessage(msg.channel, 'Usage: **`!gif`** `gif tags`');
-      return;
-    }
-    let tags = suffix.split(' ');
-    giphy(bot, msg, tags);
-  }
+  giphy
 };
