+++ lib/useful/pastebin.js
@@ -2,7 +2,9 @@ import Promise from 'bluebird';
 import nconf from 'nconf';
 import R from 'ramda';
 import _request from 'request';
+
 import sentry from './config/sentry';
+
 
 const request = Promise.promisify(_request);
