+++ lib/useful/wolfram.js
@@ -5,6 +5,7 @@ import wolframalpha from 'wolfram-alpha';
 
 import sentry from './config/sentry';
 
+
 function queryWolf(wolf, query) {
   return new Promise((resolve, reject) => {
     wolf.query(query, (err, results) => {
@@ -40,6 +41,6 @@ function wolfram(bot, msg, query) {
 }
 
 export default {
-  wolfram: wolfram,
-  wfa: wolfram
+  wfa: wolfram,
+  wolfram
 };
