+++ lib/useful/youtube.js
@@ -1,6 +1,7 @@
 import nconf from 'nconf';
 import YoutubeNode from 'youtube-node';
 
+
 const youtube = new YoutubeNode();
 youtube.setKey(nconf.get('YOUTUBE_API_KEY'));
 youtube.addParam('type', 'video,playlist');
